moced sparsevector to a header file
