+++ src/serum-decode.cpp
@@ -6,15 +6,13 @@
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
-#include <unordered_map>
-#include <vector>
 #include <chrono>
 #include <filesystem>
 #include <algorithm>
 #include <optional>
-#include <type_traits>
 
 #include "serum-version.h"
+#include "sparse-vector.h"
 
 #if defined(__APPLE__)
 #include <TargetConditionals.h>
@@ -54,63 +52,6 @@ const int pathbuflen = 4096;
 
 const uint32_t MAX_NUMBER_FRAMES = 0x7fffffff;
 const uint32_t IDENTIFY_SAME_FRAME = 0xfffffffe;
-
-template<typename T>
-class SparseVector {
-  static_assert(std::is_trivial<T>::value && std::is_standard_layout<T>::value,
-                "SparseVector only supports trivial types like uint8_t or uint16_t");
-
- protected:
-  std::unordered_map<uint32_t, std::vector<T>> data;
-  std::vector<T> noData;
-
- public:
-  SparseVector(T noDataSignature) {
-	noData.resize(1, noDataSignature);
-  }
-
-  // Access data for a frame (returns pointer to zeros if not found)
-  T* operator[](const uint32_t frame) {
-    auto it = data.find(frame);
-    if (it != data.end()) return it->second.data();
-    return noData.data();
-  }
-
-  // Load data (only store if non-zero)
-  void my_fread(size_t elementCount, uint32_t nframes, FILE* stream) {
-    size_t blockSize = elementCount * sizeof(T);
-
-    if (noData.size() < elementCount) noData.resize(elementCount, noData[0]);
-
-	std::vector<T> tmp(elementCount);
-
-    for (uint32_t i = 0; i < nframes; ++i) {
-      if (1 != fread(tmp.data(), blockSize, 1, stream)) {
-		// Error reading file
-		exit(1);
-	  }
-
-      if (memcmp(tmp.data(), noData.data(), noData.size() * sizeof(T)) != 0)
-        data[i] = tmp;
-    }
-  }
-
-  void set(uint32_t frame, const T* values, size_t elementCount) {
-	auto it = data.find(frame);
-	if (it != data.end()) {
-	  it->second.resize(elementCount);
-	  memcpy(it->second.data(), values, elementCount * sizeof(T));
-	} else {
-	  std::vector<T> tmp(values, values + elementCount);
-	  data[frame] = tmp;
-	}
-  }
-
-  // Clear all stored frames
-  void clear() {
-    data.clear();
-  }
-};
 
 // header
 char rname[64];
